U110, U202, U504, U505, U506, U115
U110: As a player, I want different animation for the character.
U202: As a player, I want different types of enemies, so I can play the game with different difficulties.
(NEW)U504: As a player, I want to have collectables so I can collect them.
(NEW)U505: As a player, I want to see the spaceship so I know where should I go/start.
(NEW)U115: As a player, I want to see animation for injuring and dying so I know the character is injure or dead.